refactor(ffi): implement PR-0411A api modularization
