+++ crates/lazynote_ffi/src/api/tests/mod.rs
@@ -1,0 +1,14 @@
+pub(super) use super::calendar::*;
+pub(super) use super::creation::*;
+pub(super) use super::entry::*;
+pub(super) use super::notes::*;
+pub(super) use super::query::*;
+pub(super) use super::tasks::*;
+pub(super) use super::test_support::*;
+pub(super) use super::workspace::*;
+pub(super) use super::*;
+
+mod entry_notes_tasks_calendar;
+mod legacy;
+mod query_and_creation;
+mod workspace;
