+++ docs/releases/v0.4/v0.4-kickoff.md
@@ -200,7 +200,7 @@ Kickoff 审计识别出两条在原始 spec 中未被预见的主要工作流，
 | **Workspace** | PR-0409 | Merged | Scoped Query Repository | `designated_folders` 作用域查询，系统节点约束 | PR-0408 |
 | **Workspace** | PR-0410 | Merged | Tree Service + Creation Service | 单根化 service 层，创建路由优先级，S7 延期：batch cancel hook | PR-0409 |
 | **Workspace** | PR-0411 | Merged | Guard FFI | `atom_create`/`atom_get`/`atom_update_content` 统一接口，旧接口兼容壳 | PR-0410 |
-| **Workspace Cleanup** | PR-0411A | Draft | FFI Structural Cleanup | `api.rs` 模块化拆分，保持 public FFI contract 不变 | PR-0411（merged 后，cleanup-only） |
+| **Workspace Cleanup** | PR-0411A | Merged | FFI Structural Cleanup | `api.rs` 模块化拆分，保持 public FFI contract 不变 | PR-0411（merged 后，cleanup-only） |
 | **Workspace** | PR-0412 | Draft | Flutter Core 收敛 | `WorkspaceTreeService` 单根化对齐，`atom_create` 消费 | PR-0411 |
 | **Workspace** | PR-0413 | Draft | Flutter Features 收敛 | Explorer 去 synthetic uncategorized，Tasks/Calendar 面板同源分组 | PR-0412 |
 | **Workspace Cleanup** | PR-0413A | Draft | Workspace Structural Cleanup | `tree_repo.rs` 与 `workspace_tree.rs` 拆分，保持行为语义不变 | PR-0413（merged 后，cleanup-only） |
